fix: move basecolors to ahve it in the dist
